+++ test/features/trips/presentation/cubits/trip_cubit_test.mocks.dart
@@ -397,6 +397,14 @@ class MockCategoryRepository extends _i1.Mock
           as _i5.Future<_i3.Category?>);
 
   @override
+  _i5.Future<List<_i3.Category>> getCategoriesByIds(List<String>? ids) =>
+      (super.noSuchMethod(
+            Invocation.method(#getCategoriesByIds, [ids]),
+            returnValue: _i5.Future<List<_i3.Category>>.value(<_i3.Category>[]),
+          )
+          as _i5.Future<List<_i3.Category>>);
+
+  @override
   _i5.Future<_i3.Category> createCategory({
     required String? name,
     String? icon = 'label',