+++ test/integration/category_customization_flow_test.mocks.dart
@@ -8,6 +8,8 @@ import 'dart:async' as _i4;
 import 'package:expense_tracker/core/models/category_customization.dart' as _i5;
 import 'package:expense_tracker/core/repositories/category_customization_repository.dart'
     as _i3;
+import 'package:expense_tracker/features/categories/domain/models/category.dart'
+    as _i10;
 import 'package:expense_tracker/features/categories/presentation/cubit/category_cubit.dart'
     as _i9;
 import 'package:expense_tracker/features/categories/presentation/cubit/category_state.dart'
@@ -16,7 +18,7 @@ import 'package:expense_tracker/features/trips/domain/models/activity_log.dart'
     as _i7;
 import 'package:expense_tracker/features/trips/domain/repositories/activity_log_repository.dart'
     as _i6;
-import 'package:flutter_bloc/flutter_bloc.dart' as _i10;
+import 'package:flutter_bloc/flutter_bloc.dart' as _i11;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:mockito/src/dummies.dart' as _i8;
 
@@ -219,6 +221,20 @@ class MockCategoryCubit extends _i1.Mock implements _i9.CategoryCubit {
           as _i4.Future<void>);
 
   @override
+  _i4.Future<void> loadCategoriesByIds(List<String>? ids) =>
+      (super.noSuchMethod(
+            Invocation.method(#loadCategoriesByIds, [ids]),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
+
+  @override
+  _i10.Category? getCategoryById(String? categoryId) =>
+      (super.noSuchMethod(Invocation.method(#getCategoryById, [categoryId]))
+          as _i10.Category?);
+
+  @override
   _i4.Future<void> close() =>
       (super.noSuchMethod(
             Invocation.method(#close, []),
@@ -234,7 +250,7 @@ class MockCategoryCubit extends _i1.Mock implements _i9.CategoryCubit {
   );
 
   @override
-  void onChange(_i10.Change<_i2.CategoryState>? change) => super.noSuchMethod(
+  void onChange(_i11.Change<_i2.CategoryState>? change) => super.noSuchMethod(
     Invocation.method(#onChange, [change]),
     returnValueForMissingStub: null,
   );